index: improve and fix index, add tags
Fix internal links title and url. It changes the specs to be more
exhaustive and cable everything. We ensure that entries are generated
with the right title and url if not found.

We also add the tags in each entries, so we can query on it mandatory,
optional, and forbidden tags to find pages by topic.